backend/requirements.txt
Flask
flask-cors
pymongo
dnspython
Flask-JWT-Extended
python-dotenv
